pyproject.toml
[tool.poetry]
name = "wundernut-vol12"
version = "0.1.0"
description = ""
authors = ["Antti-Jussi Inkinen <aj.inkinen@gmail.com>"]

[tool.poetry.dependencies]
python = "3.10.6"
Pillow = "^9.2.0"
easyocr = "^1.6.0"
wordsegment = "^1.3.1"

[tool.poetry.group.dev.dependencies]
flake8 = "^5.0.4"
mypy = "^0.971"

[tool.mypy]
plugins = 'numpy.typing.mypy_plugin'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
